yoyiyi

+++ app/src/main/java/com/yoyiyi/soleil/bean/bangumi/BangumiDetailComment.java
@@ -320,7 +320,7 @@ public class BangumiDetailComment {
                     public int pid;
                     public String name;
                     public String image;
-                    public int expire;
+                    public long expire;
                 }
 
                 public static class NameplateBeanX {
